+++ inc/form_element_array.php
@@ -280,6 +280,13 @@ class form_element_array extends form_element {
   function show_element($document) {
     $div=parent::show_element($document);
 
+    $createable = ((!array_key_exists("createable", $this->def)) || ($this->def['createable'] == true)) ? "createable": "not_createable";
+
+    if ($this->tr) {
+      $this->tr->setAttribute('class', $this->tr->getAttribute('class') . " $createable");
+    }
+    $this->dom->setAttribute('class', $this->dom->getAttribute('class') . " $createable");
+
     foreach($this->elements as $k=>$element) {
       $part_div=$this->show_element_part($k, $element, $document);
       $div->appendChild($part_div);
@@ -289,18 +296,20 @@ class form_element_array extends form_element {
     $el_div->setAttribute("class", "form_element_array_actions");
     $div->appendChild($el_div);
 
-    $input=$document->createElement("input");
-    $input->setAttribute("type", "submit");
-    $input->setAttribute("name", "{$this->options['var_name']}[__new]");
-    if(array_key_exists("button:add_element", $this->def)) {
-      if(is_array($this->def['button:add_element']))
-        $input->setAttribute("value", lang($this->def['button:add_element']));
+    if ($createable === 'createable') {
+      $input=$document->createElement("input");
+      $input->setAttribute("type", "submit");
+      $input->setAttribute("name", "{$this->options['var_name']}[__new]");
+      if(array_key_exists("button:add_element", $this->def)) {
+        if(is_array($this->def['button:add_element']))
+          $input->setAttribute("value", lang($this->def['button:add_element']));
+        else
+          $input->setAttribute("value", $this->def['button:add_element']);
+      }
       else
-        $input->setAttribute("value", $this->def['button:add_element']);
+        $input->setAttribute("value", lang("form:add_element"));
+      $el_div->appendChild($input);
     }
-    else
-      $input->setAttribute("value", lang("form:add_element"));
-    $el_div->appendChild($input);
 
     if(isset($this->def['max']) && (sizeof($this->elements) >= $this->def['max'])) {
       $input->setAttribute("class", "reached_max");
@@ -362,4 +371,40 @@ class form_element_array extends form_element {
       $element->refresh($force);
     }
   }
-}
+
+  function check_count (&$errors, $param) {
+    $op = $param[0];
+    $value = $param[1];
+    $count = sizeof($this->get_data());
+    $result;
+
+    switch ($op) {
+      case '>=':
+        $result = $count >= $value;
+        break;
+      case '>':
+        $result = $count > $value;
+        break;
+      case '<':
+        $result = $count < $value;
+        break;
+      case '<=':
+        $result = $count <= $value;
+        break;
+      case '==':
+        $result = $count == $value;
+        break;
+      case '!=':
+        $result = $count != $value;
+        break;
+      default:
+        $errors[] = 'invalid operator';
+    }
+
+    if(!$result) {
+      if (sizeof($param) < 3)
+        $errors[] = lang('form:require_value');
+      else
+        $errors[] = $param[2];
+    }
+  }}