+++ inc/form_element_autocomplete.js
@@ -24,6 +24,10 @@ form_element_autocomplete.prototype.connect=function(dom_parent) {
     this.set_data(value);
 }
 
+form_element_autocomplete.prototype.focus = function() {
+  this.dom_element.focus()
+}
+
 form_element_autocomplete.prototype.create_element=function(div) {
   var cls="form_orig";
   if(this.is_modified())
@@ -35,7 +39,9 @@ form_element_autocomplete.prototype.create_element=function(div) {
   if(this.def.html_attributes)
     for(var i in this.def.html_attributes)
       input.setAttribute(i, this.def.html_attributes[i]);
-  input.name=this.options.var_name;
+  if (this.options.var_name) {
+    input.name=this.options.var_name;
+  }
 
   div.appendChild(input);