+++ inc/form_element_boolean.js
@@ -13,6 +13,10 @@ form_element_boolean.prototype.connect=function(dom_parent) {
   this.dom_element.onchange=this.notify_change.bind(this);
 }
 
+form_element_boolean.prototype.focus = function() {
+  this.dom_element.focus()
+}
+
 form_element_boolean.prototype.show_element=function() {
   var div=this.parent("form_element_boolean").show_element.call(this);
 
@@ -31,8 +35,13 @@ form_element_boolean.prototype.show_element=function() {
     for(var i in this.def.html_attributes)
       input.setAttribute(i, this.def.html_attributes[i]);
 
-  input.name=this.options.var_name;
-  input.id=this.options.var_name;
+  if (this.options.var_name) {
+    input.name=this.options.var_name;
+    input.id=this.options.var_name;
+  } else {
+    input.id = Math.random()
+  }
+
   input.value="on";
   if((this.data==="on")||(this.data===true))
     input.setAttribute("checked", "checked");
@@ -42,7 +51,7 @@ form_element_boolean.prototype.show_element=function() {
   this.dom_element.onchange=this.notify_change.bind(this);
 
   var label=document.createElement("label");
-  label.setAttribute("for", this.options.var_name);
+  label.setAttribute("for", input.id);
   span.appendChild(label);
 
   return div;