+++ inc/form_element_form_chooser.php
@@ -6,7 +6,7 @@ class form_element_form_chooser extends form_element {
     parent::__construct($id, $def, $options, $form_parent);
     $this->changed_count=false;
 
-    $this->build_form();
+    $this->build_form(true);
   }
 
   function is_complete() {
@@ -51,6 +51,14 @@ class form_element_form_chooser extends form_element {
 
       if($element->include_data())
 	$data[$k] = $d;
+    }
+
+    foreach($this->available_elements as $k=>$element) {
+      if (!array_key_exists($k, $this->elements)) {
+        if (array_key_exists('non_used_value', $element->def)) {
+          $data[$k] = $element->def['non_used_value'];
+        }
+      }
     }
 
     if (array_key_exists('result_keep_order', $this->def) && $this->def['result_keep_order']) {
@@ -236,13 +244,15 @@ class form_element_form_chooser extends form_element {
     return $ret;
   }
 
-  function build_form() {
+  function build_form($show_default=false) {
     $this->elements=array();
     $this->available_elements=array();
 
     if (!$this->data) {
       $this->data = array();
     }
+
+    $this->def['def'] = weight_sort($this->def['def']);
 
     foreach ($this->def['def'] as $k => $def) {
       $element_def = $def;
@@ -254,6 +264,10 @@ class form_element_form_chooser extends form_element {
       if (class_exists($element_class)) {
         $this->available_elements[$k] = new $element_class($element_id, $element_def, $element_options, $this);
       } else {
+      }
+
+      if ($show_default && array_key_exists('show_default', $def) && $def['show_default']) {
+        $this->add_element($k);
       }
     }
   }